backend/requirements.txt
# FastAPI核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
supabase==2.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# 天勤TqSDK
tqsdk==3.6.2

# 数据模型
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP客户端
httpx==0.26.0
requests==2.31.0

# WebSocket
websockets==12.0

# 工具
python-dotenv==1.0.0
python-dateutil==2.8.2
psutil==5.9.8

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
